separate io::Error from crate::Error
