+++ libs/gesha-core/src/io/mod.rs
@@ -1,3 +1,6 @@
+mod error;
+pub use error::Error;
+
 mod reader;
 pub use reader::Reader;
